Changing the way of List::display() implementation - other stuff I dont remember

+++ include/DisplayBoundary.h
@@ -34,7 +34,7 @@ public:
 	DisplayPosition getBegin();
 	DisplayPosition getEnd();
 
-    virtual int16_t getSerializedSize();
+    virtual int getSerializedSize();
 	virtual int serialize(void* const data);
 	virtual int deserialize(void const * const data);
