+++ include/DisplayPosition.h
@@ -17,7 +17,7 @@ class DisplayPosition : public Serializeable
         void setXPosition(int xpos);
         void setYPosition(int ypos);
 
-        virtual int16_t getSerializedSize();
+        virtual int getSerializedSize();
     	virtual int serialize(void* const data);
     	virtual int deserialize(void const * const data);
